n+1 problem fix

+++ app/Http/Controllers/CategoryController.php
@@ -4,12 +4,14 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Category;
+use Illuminate\Support\Facades\Auth;
+
 
 class CategoryController extends Controller
 {
     public function index()
     {
-        $categories = Category::withCount('todos')->get();
+        $categories = Category::with('todos')->where('user_id', Auth::id())->get();
         return view('category.index', compact('categories'));
     }
 
@@ -20,26 +22,50 @@ class CategoryController extends Controller
 
     public function store(Request $request)
     {
-        $request->validate(['name' => 'required|string|max:255']);
-        Category::create(['name' => ucfirst($request->name)]);
-        return redirect()->route('category.index')->with('success', 'Category created.');
+        $request->validate([
+        'title' => 'required|string|max:255',
+        ]);
+
+        Category::create([
+            'title' => $request->title,
+            'user_id' => auth()->id(),
+        ]);
+
+        return redirect()->route('category.index')->with('success', 'Category created successfully!');
     }
 
     public function edit(Category $category)
     {
-        return view('category.edit', compact('category'));
+        if (Auth::id() == $category->user_id) {
+            return view('category.edit', compact('category'));
+        } else {
+            return redirect()->route('category.index')->with('danger', 'You are not authorized to edit this category!');
+        }
     }
 
     public function update(Request $request, Category $category)
     {
-        $request->validate(['name' => 'required|string|max:255']);
-        $category->update(['name' => ucfirst($request->name)]);
-        return redirect()->route('category.index')->with('success', 'Category updated.');
+
+        $request->validate([
+            'title' => 'required|string|max:255',
+        ]);
+
+        $category->update([
+            'title' => $request->title,
+        ]);
+
+        return redirect()->route('category.index')->with('success', 'Kategori berhasil diperbarui!');
     }
+
+    
 
     public function destroy(Category $category)
     {
-        $category->delete();
-        return redirect()->route('category.index')->with('success', 'Category deleted.');
+       if (Auth::id() == $category->user_id) {
+            $category->delete();
+            return redirect()->route('category.index')->with('success', 'Category deleted successfully!');
+        } else {
+            return redirect()->route('category.index')->with('danger', 'You are not authorized to delete this category!');
+        }
     }
 }
